Move away from old shape package
