--- util/shape/doc.go
@@ -1,3 +1,0 @@
-// Package shape provides types and functions for defining 3D shapes and
-// performing intersection (collision) checks between them.
-package shape
